Move config around and start using version catalogs
