+++ gradle/libs.versions.toml
@@ -1,0 +1,3 @@
+
+[plugins]
+j52j = { id = "dev.kikugie.j52j", version = "2.0"}
